+++ app/(auth)/create-new-account.tsx
@@ -101,7 +101,7 @@ const SignUpScreen = () => {
           Already have an account?{" "}
           <Text
             style={styles.loginLink}
-            onPress={() => router.navigate("/(auth)/login")}
+            onPress={() => router.replace("/(auth)/login")}
           >
             Login
           </Text>
