+++ app/(auth)/login.tsx
@@ -68,7 +68,7 @@ const LoginScreen = () => {
           Don't have an account?{" "}
           <Text
             style={styles.footerLink}
-            onPress={() => router.navigate("/create-new-account")}
+            onPress={() => router.replace("/create-new-account")}
           >
             Sign Up
           </Text>
@@ -76,7 +76,7 @@ const LoginScreen = () => {
 
         <Text
           style={styles.footerLink}
-          onPress={() => router.navigate("/forget-password")}
+          onPress={() => router.replace("/forgot-password")}
         >
           Forgot Password?
         </Text>
