--- app/(auth)/signup.tsx
@@ -1,34 +1,0 @@
-import { Button } from "react-native";
-import { Text, View } from "@/components/theme/Themed";
-import { useAuthContext } from "@/contexts";
-import { DUMMY_MANAGER_CREDENTIALS } from "@/constants/Manager";
-
-const Signup = () => {
-  const {
-    signUp,
-  } = useAuthContext();
-
-  const handleSignUp = () => {
-    signUp(
-      DUMMY_MANAGER_CREDENTIALS.name,
-      DUMMY_MANAGER_CREDENTIALS.email,
-      DUMMY_MANAGER_CREDENTIALS.password
-    );
-  };
-
-  return (
-    <View
-      style={{
-        flex: 1,
-        justifyContent: "center",
-        alignItems: "center",
-        gap: 20,
-      }}
-    >
-      <Text>Signup</Text>
-      <Button title="Signup" onPress={handleSignUp} />
-    </View>
-  );
-};
-
-export default Signup;
